feat: seperate Author and Viewer from ImgLayers.
